requirements.txt
alembic==1.15.2
blinker==1.9.0
cffi==1.17.1
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
email-validator==2.0.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
